Changes necessary for making the project work on Windows, with MSVC.
Tweaked the CMake configuration to make it work correctly with MSVC,
and made some small changes in the code. Mainly to make certain type
conversions explicit.

+++ tests/accelerator/cuda/common/execute_cuda_test.cuh
@@ -42,11 +42,12 @@ void execute_cuda_test(std::size_t arraySizes, ARGS... args) {
   // If the arrays are not even this large, then reduce the value to the
   // size of the arrays.
   if (arraySizes < nThreadsPerBlock) {
-    nThreadsPerBlock = arraySizes;
+    nThreadsPerBlock = static_cast<int>(arraySizes);
   }
 
   // Launch the test on the device.
-  const int nBlocks = ((arraySizes + nThreadsPerBlock - 1) / nThreadsPerBlock);
+  const int nBlocks =
+      static_cast<int>((arraySizes + nThreadsPerBlock - 1) / nThreadsPerBlock);
   cudaTestKernel<FUNCTOR><<<nBlocks, nThreadsPerBlock>>>(arraySizes, args...);
 
   // Check whether it succeeded to run.
